Collections working as expected. Added separate collection route file because the post routes was getting too big to maintain
