+++ app/main/posts/collections/collections.html
@@ -1,8 +1,5 @@
 <div>
     <layout-class layout="{{layout}}"></layout-class>
     <collection-mode-context  ng-if="layout === 'a'"></collection-mode-context>
-
-    <main role="main">
-        <div ui-view></div>
-    </main>
+    <div ui-view="collectionData"></div>
 </div>
