+++ app/main/posts/views/main.html
@@ -28,6 +28,7 @@
         <!--@uirouter-refactor ask Anna about this <ng-transclude ng-if="currentView !== 'data'"></ng-transclude>-->
         <!-- todo: toolbox-->
         <div ui-view noanimation="true"></div>
+        <div ui-view="collectionsAll" noanimation="true"></div>
         <ush-logo></ush-logo>
     </main>
 </div>
